Chapter09/automation_bundle_and_cicd_project/requirements-test.txt
pytest>=7.0.0
databricks-connect>=15.0.0
pytest-cov>=4.0.0
